refactor for atomic design
