+++ src/components/atoms/Links/HeartLink/assets/icon--heart.png
