--- src/components/molecules/NavBar/NavLinksData.js
@@ -1,11 +1,0 @@
-import homeLogo from './assets/icon--home.png';
-import browseLogo from './assets/icon--browse.png';
-import profileLogo from './assets/icon--profile.png';
-
-const Links = [
-  { id: '0', text: 'home', image: homeLogo, link: '/dashboard' },
-  { id: '1', text: 'browse', image: browseLogo, link: '/browse' },
-  { id: '2', text: 'profile', image: profileLogo, link: '/profile' },
-];
-
-export default Links;
